requirements.txt
# Azure Functions Core
azure-functions

# Security & Validation
python-dotenv==1.0.0
email-validator==2.2.0
bleach==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring & Logging
azure-monitor-opentelemetry==1.2.0

# Database & Storage
azure-cosmos==4.5.1
azure-data-tables==12.4.4

# Caching & Performance
redis==5.0.1

# Real-time Features
websockets==12.0

# Code Execution & Security
psutil==5.9.8

# Additional Utilities
uuid
asyncio

# AI and Machine Learning
numpy==1.24.3
scikit-learn==1.3.0

# Enhanced Security
psutil==5.9.8

# Email functionality is built into Python standard library
# No additional dependencies needed for smtplib, email.mime
